Separate precompile tests into their own module
